gallery list pull to refresh, load more 처리 / thumbnail size 조절

+++ core/src/main/java/com/carrot/gallery/core/apis/ImageApis.kt
@@ -15,6 +15,6 @@ interface ImageApis {
     suspend fun getImages(@Query("page") page: Int, @Query("limit") limit: Int): List<Image>
 
     @GET("/id/{id}/info")
-    suspend fun getImage(@Path("id") id: Int): Image
+    suspend fun getImage(@Path("id") id: Long): Image
 
 }
